worker/requirements.txt
celery
redis
pillow
google-cloud-storage
psycopg2-binary
sqlalchemy
python-dotenv